fix command query

+++ deleteUser/index.js
@@ -3,8 +3,13 @@ const CONSTANTS = require('./constant');
 exports.handler = async (query) => {
     try {
         const userId = query.arguments.input.userId;
+        if(userId)
+        {
         await DB.deleteUser(userId);
         return { message: 'Successfully Deleted' }
+        } else {
+            return { message: 'Invalid UserId' }
+        }
     } catch (error) {
         throw error;
     }
